+++ CHANGELOG.md
@@ -62,6 +62,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Ability to rearrange order of attachments when uploading
 - Enabled users to zoom and pan images in media viewer with mouse and touch
 - Added frontend ui for account migration
+- Implemented remote interaction with statuses
 
 
 ## [2.4.2] - 2022-01-09